backend/requirements.txt
fastapi
uvicorn[standard]
httpx
qrcode[pil]
pillow
python-dotenv
sse-starlette
